python/python_mouse_click/requirements.txt
enum34==1.1.6
pynput==1.1
pyobjc-core==3.1.1
pyobjc-framework-Cocoa==3.1.1
pyobjc-framework-Quartz==3.1.1
six==1.10.0
